+++ source/include/memory.h
@@ -37,22 +37,20 @@
 #include <config.h>
 #endif
 
-#include <stdlib.h> /* for size_t */
-
 #include "include/float.h"
 #include "include/ordinals.h" /* for FLAC__bool */
 
 /* Returns the unaligned address returned by malloc.
  * Use FLAC_FREE() on this address to deallocate.
  */
-void *FLAC__memory_alloc_aligned(size_t bytes, void **aligned_address);
-FLAC__bool FLAC__memory_alloc_aligned_int32_array(size_t elements, FLAC__int32 **unaligned_pointer, FLAC__int32 **aligned_pointer);
-FLAC__bool FLAC__memory_alloc_aligned_uint32_array(size_t elements, FLAC__uint32 **unaligned_pointer, FLAC__uint32 **aligned_pointer);
-FLAC__bool FLAC__memory_alloc_aligned_uint64_array(size_t elements, FLAC__uint64 **unaligned_pointer, FLAC__uint64 **aligned_pointer);
-FLAC__bool FLAC__memory_alloc_aligned_unsigned_array(size_t elements, uint32_t **unaligned_pointer, uint32_t **aligned_pointer);
+void *FLAC__memory_alloc_aligned(uint32_t bytes, void **aligned_address);
+FLAC__bool FLAC__memory_alloc_aligned_int32_array(uint32_t elements, FLAC__int32 **unaligned_pointer, FLAC__int32 **aligned_pointer);
+FLAC__bool FLAC__memory_alloc_aligned_uint32_array(uint32_t elements, FLAC__uint32 **unaligned_pointer, FLAC__uint32 **aligned_pointer);
+FLAC__bool FLAC__memory_alloc_aligned_uint64_array(uint32_t elements, FLAC__uint64 **unaligned_pointer, FLAC__uint64 **aligned_pointer);
+FLAC__bool FLAC__memory_alloc_aligned_unsigned_array(uint32_t elements, uint32_t **unaligned_pointer, uint32_t **aligned_pointer);
 #ifndef FLAC__INTEGER_ONLY_LIBRARY
-FLAC__bool FLAC__memory_alloc_aligned_real_array(size_t elements, FLAC__real **unaligned_pointer, FLAC__real **aligned_pointer);
+FLAC__bool FLAC__memory_alloc_aligned_real_array(uint32_t elements, FLAC__real **unaligned_pointer, FLAC__real **aligned_pointer);
 #endif
-void *safe_malloc_mul_2op_p(size_t size1, size_t size2);
+void *safe_malloc_mul_2op_p(uint32_t size1, uint32_t size2);
 
 #endif
