+++ source/memory.c
@@ -43,7 +43,7 @@
 #include "include/compat.h"
 #include "include/alloc.h"
 
-void *FLAC__memory_alloc_aligned(size_t bytes, void **aligned_address)
+void *FLAC__memory_alloc_aligned(uint32_t bytes, void **aligned_address)
 {
 	void *x;
 
@@ -60,7 +60,7 @@ void *FLAC__memory_alloc_aligned(size_t bytes, void **aligned_address)
 	return x;
 }
 
-FLAC__bool FLAC__memory_alloc_aligned_int32_array(size_t elements, FLAC__int32 **unaligned_pointer, FLAC__int32 **aligned_pointer)
+FLAC__bool FLAC__memory_alloc_aligned_int32_array(uint32_t elements, FLAC__int32 **unaligned_pointer, FLAC__int32 **aligned_pointer)
 {
 	FLAC__int32 *pu;	 /* unaligned pointer */
 	union {				 /* union needed to comply with C99 pointer aliasing rules */
@@ -91,7 +91,7 @@ FLAC__bool FLAC__memory_alloc_aligned_int32_array(size_t elements, FLAC__int32 *
 	}
 }
 
-FLAC__bool FLAC__memory_alloc_aligned_uint32_array(size_t elements, FLAC__uint32 **unaligned_pointer, FLAC__uint32 **aligned_pointer)
+FLAC__bool FLAC__memory_alloc_aligned_uint32_array(uint32_t elements, FLAC__uint32 **unaligned_pointer, FLAC__uint32 **aligned_pointer)
 {
 	FLAC__uint32 *pu;	 /* unaligned pointer */
 	union {				  /* union needed to comply with C99 pointer aliasing rules */
@@ -122,7 +122,7 @@ FLAC__bool FLAC__memory_alloc_aligned_uint32_array(size_t elements, FLAC__uint32
 	}
 }
 
-FLAC__bool FLAC__memory_alloc_aligned_uint64_array(size_t elements, FLAC__uint64 **unaligned_pointer, FLAC__uint64 **aligned_pointer)
+FLAC__bool FLAC__memory_alloc_aligned_uint64_array(uint32_t elements, FLAC__uint64 **unaligned_pointer, FLAC__uint64 **aligned_pointer)
 {
 	FLAC__uint64 *pu;	 /* unaligned pointer */
 	union {				  /* union needed to comply with C99 pointer aliasing rules */
@@ -153,7 +153,7 @@ FLAC__bool FLAC__memory_alloc_aligned_uint64_array(size_t elements, FLAC__uint64
 	}
 }
 
-FLAC__bool FLAC__memory_alloc_aligned_unsigned_array(size_t elements, uint32_t **unaligned_pointer, uint32_t **aligned_pointer)
+FLAC__bool FLAC__memory_alloc_aligned_unsigned_array(uint32_t elements, uint32_t **unaligned_pointer, uint32_t **aligned_pointer)
 {
 	uint32_t *pu;	 /* unaligned pointer */
 	union {			  /* union needed to comply with C99 pointer aliasing rules */
@@ -186,7 +186,7 @@ FLAC__bool FLAC__memory_alloc_aligned_unsigned_array(size_t elements, uint32_t *
 
 #ifndef FLAC__INTEGER_ONLY_LIBRARY
 
-FLAC__bool FLAC__memory_alloc_aligned_real_array(size_t elements, FLAC__real **unaligned_pointer, FLAC__real **aligned_pointer)
+FLAC__bool FLAC__memory_alloc_aligned_real_array(uint32_t elements, FLAC__real **unaligned_pointer, FLAC__real **aligned_pointer)
 {
 	FLAC__real *pu;		/* unaligned pointer */
 	union {				/* union needed to comply with C99 pointer aliasing rules */
@@ -219,7 +219,7 @@ FLAC__bool FLAC__memory_alloc_aligned_real_array(size_t elements, FLAC__real **u
 
 #endif
 
-void *safe_malloc_mul_2op_p(size_t size1, size_t size2)
+void *safe_malloc_mul_2op_p(uint32_t size1, uint32_t size2)
 {
 	if (!size1 || !size2)
 	{
